+++ smart-user-api/src/main/java/com/smartitengineering/user/domain/UserGroup.java
@@ -15,7 +15,7 @@ import org.codehaus.jackson.annotate.JsonIgnore;
  *
  * @author russel
  */
-public class UserGroup extends AbstractGenericPersistentDTO<UserGroup, Long, Integer> {
+public class UserGroup extends AbstractGenericPersistentDTO<UserGroup, Long, Long> {
 
   private String name;
   private Set<User> users;
